Now our class has some news methods to help us
Now all the previous 'search_top' methods have been moved to class methods. The reason for this change is to re-use the same methods for both 'search_top' and 'search_all' methods.

Since the first app and the rest are in different divs, and each div has its own classes, we need to use a different approach to get the first app and the rest of the apps. The code is self-explanatory.